Moved SrsButton into its own file, fixed border-radius bug and now always displaying total number of items in each stage
